+++ tests/CMakeLists.txt
@@ -1,7 +1,7 @@
 
 add_executable(all_tests
         all_tests.cpp
-        test_model.cpp ctrl_tests.cpp test_jwt.cpp pkcs7_tests.cpp)
+        test_ziti_model.cpp ctrl_tests.cpp test_jwt.cpp pkcs7_tests.cpp)
 set_property(TARGET all_tests PROPERTY CXX_STANDARD 11)
 
 target_include_directories(all_tests
